chatting 스키마 모델 분리
