+++ app/schemas/__init__.py
@@ -2,3 +2,4 @@ from .quiz import *
 from .user import *
 from .scoreboard import *
 from .community import *
+from .chatting import *
